Implement a "New item" handler on server
The handler is at endpoint http://localhost:8080/api/newItem. It gets
the form value with key "Title" and inserts that value into
the database as the title of a new todo item. The handler passes a
unit test: the test sends an http request to add a new item.
This unit test lives in the new file main_test.go.