Implement deleteItem handler function
The unit test for deleteItem should (but does not yet) add an item, check
that the item is there, then delete the item, then check that the item is gone.
Currently all it does is make the api call to delete an item, leaving it up to
the user to check manually that it's been deleted.