Add a dummy "find item" endpoint
It currently prints out the number of matching rows it found to the console.
It wants to be a helper function for the first good test, which will add and
delete one item from the database, checking before, in between, and after these
operations for the presence of the item.