Add an error page which will load if the fetch command fails
This commit adds another state variable to the TodoList component, a boolean
which indicates if an error occurred or not. Currently this state is only
set to true if the fetch call in componentDidMount fails. It is a simple
boolean and does not know the difference between any kinds of errors. It could
be extended to print the error message, for example.