Make code pass TestHandlers/Print test after MySQL is installed via script
The code also passes all the frontend and backend tests on my local machine,
where MySQL and Go are already installed. When run on an Amazon ec2 instance,
the code was adding a newline after reading the password from password.txt via
ioutil.Readline. I added code to trim the newline if it was present.