Complete the working prototype for AWS Linux
The install script now works on an AWS Linux instance. The steps to get it to work on a brand
new instance are
--Run "sudo yum groupinstall "Development Tools"" to install git and gcc, among other things
--Create ~/go/src/github.com/ericauld directory
--Clone the repo
--Run the install script
--Source the bash profile
At that point, the tests in the testbackend file pass.

The script still has bugs when it prompts the user for a password; the test is all jumbled up
and it's not properly displaying newlines. We have added testfrontend and testbackend to the repo
(removed them from .gitignore), but are dangerous because at their conclusion they simply kill
all processes running on :8080. If a user has processes running on this port, it may cause
problems for them.